+++ CHANGELOG.md
@@ -18,6 +18,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Updated all dependencies
 
+### Changed
+
+ - Notifications for group messages now display the name of the group instead of just the name of the sender.
+
 ## [0.6.0] - 2023-01-13
 
 ### Added