+++ CHANGELOG.md
@@ -16,13 +16,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Showing profile names
 - Copy action in message menu
 
+### Changed
+
+- Notifications for group messages now display the name of the group instead of just the name of the sender.
+- Messages must now be successfully sent before displaying in a chat.
+
 ### Chores
 
 - Updated all dependencies
 
-### Changed
+### Fixed
 
- - Notifications for group messages now display the name of the group instead of just the name of the sender.
+- Fixed not setting revision for groups.
 
 ## [0.6.0] - 2023-01-13