+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.7.0] - 2023-03-29
+
 ### BREAKING
 
 - Due to a new storage for data, a relink is required on first startup
@@ -292,7 +294,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed cargo clippy.
 - Updated presage to official repository.
 
-[Unreleased]: https://gitlab.com/Schmiddiii/flare/-/compare/0.6.0...master
+[Unreleased]: https://gitlab.com/Schmiddiii/flare/-/compare/0.7.0...master
+[0.7.0]: https://gitlab.com/Schmiddiii/flare/-/compare/0.6.0...0.7.0
 [0.6.0]: https://gitlab.com/Schmiddiii/flare/-/compare/0.5.6...0.6.0
 [0.5.6]: https://gitlab.com/Schmiddiii/flare/-/compare/0.5.5...0.5.6
 [0.5.5]: https://gitlab.com/Schmiddiii/flare/-/compare/0.5.4...0.5.5